Stop removing libraries from local maven
While it's a valid practice to remove all files that were downloaded during
the test execution, here we don't check that files were not present
before test start and we removing files that might be used by
someone else. So let them be - we will download less during the
next execution.